added error handling for request data for the skill controller

+++ employee-skills-tracking-application/employee-skills-tracking-application/src/main/java/com/example/employeeskillstrackingapplication/SkillRestController.java
@@ -3,6 +3,7 @@ package com.example.employeeskillstrackingapplication;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -10,7 +11,10 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+
+import exceptions.InvalidDataException;
 
 @CrossOrigin
 @RestController
@@ -18,7 +22,7 @@ public class SkillRestController {
 
 	@Autowired
 	SkillService skillService;
-	
+
 	@Autowired
 	EmployeeService employeeService;
 
@@ -28,34 +32,58 @@ public class SkillRestController {
 	}
 
 	@PostMapping("/employees/{employeeId}/skills")
-	public long saveSkill(@RequestBody Skill skill,
-			@PathVariable("employeeId") int employeeId) {
+	@ResponseStatus(HttpStatus.CREATED)
+	public long saveSkill(@RequestBody Skill skill, @PathVariable("employeeId") int employeeId) {
 		Employee employee = employeeService.getEmployeeById(employeeId);
+		if (employee == null) {
+			throw new InvalidDataException("Invalid Perficient employee data sent to server.");
+		}
 		skill.setEmployee(employee);
-		skillService.saveOrUpdate(skill);
-		return skill.getSkillId();
+		Skill newSkill = skillService.save(skill);
+		if (newSkill != null) {
+			return skill.getSkillId();
+		} else {
+			throw new InvalidDataException("Invalid Perficient skill data sent to the server");
+		}
 	}
-	
+
 	@GetMapping("/employees/{employeeId}/skills/{skillId}")
-	public Skill getSkillById(@PathVariable("employeeId") int employeeId, 
-			@PathVariable("skillId") long skillId) {
+	public Skill getSkillById(@PathVariable("employeeId") int employeeId, @PathVariable("skillId") long skillId) {
+		Employee employee = employeeService.getEmployeeById(employeeId);
+		if (employee == null) {
+			throw new InvalidDataException("Invalid Perficient employee data sent to server.");
+		}
 		return skillService.getSkillById(skillId);
+
 	}
-	
+
 	@PutMapping("/employees/{employeeId}/skills/{skillId}")
-	public long updateSkill(@RequestBody Skill skill,
-			@PathVariable("employeeId") int employeeId,
+	public long updateSkill(@RequestBody Skill skill, @PathVariable("employeeId") int employeeId,
 			@PathVariable("skillId") long skillId) {
 		Employee employee = employeeService.getEmployeeById(employeeId);
+		if (employee == null) {
+			throw new InvalidDataException("Invalid Perficient employee data sent to server.");
+		}
 		skill.setEmployee(employee);
 		skill.setSkillId(skillId);
-		skillService.saveOrUpdate(skill);
-		return skill.getSkillId();
+		Skill updatedSkill = skillService.update(skill);
+		if (updatedSkill != null) {
+			return updatedSkill.getSkillId();
+		} else {
+			throw new InvalidDataException("Invalid Perficient skill data sent to the server");
+		}
 	}
-	
+
 	@DeleteMapping("/employees/{employeeId}/skills/{skillId}")
-	public void deleteEmployee(@PathVariable("skillId") Long skillId) {
+	@ResponseStatus(HttpStatus.NO_CONTENT)
+	public void deleteEmployee( @PathVariable("employeeId") int employeeId,
+			@PathVariable("skillId") Long skillId) {
+		Employee employee = employeeService.getEmployeeById(employeeId);
+		if (employee == null) {
+			throw new InvalidDataException("Invalid Perficient employee data sent to server.");
+		}
+		Skill skillToDelete = skillService.getSkillById(skillId);
 		skillService.delete(skillId);
 	}
-	
+
 }
